test(cli): move CLI config tests to a dedicated config_tests.rs
Refactor CLI configuration tests by moving them from src/cli/config.rs into a new separate file src/cli/config_tests.rs to improve organization and maintainability.

Additionally, add new Unix-only runner tests for CLI default-target execution in tests/runner_cases/default_targets.rs and include them in tests/runner_tests.rs, refining the test suite structure.
