+++ src/cli/config.rs
@@ -307,113 +307,5 @@ impl CliConfig {
 }
 
 #[cfg(test)]
-mod tests {
-    use super::*;
-    use rstest::rstest;
-    use serde_json::json;
-
-    #[rstest]
-    #[case::colour_auto(
-        ColourPolicy::parse_raw("auto").map(|value| value.to_string()),
-        "auto"
-    )]
-    #[case::colour_always(
-        ColourPolicy::parse_raw("always").map(|value| value.to_string()),
-        "always"
-    )]
-    #[case::spinner_enabled(
-        SpinnerMode::parse_raw("enabled").map(|value| value.to_string()),
-        "enabled"
-    )]
-    #[case::spinner_disabled(
-        SpinnerMode::parse_raw("disabled").map(|value| value.to_string()),
-        "disabled"
-    )]
-    #[case::output_human(
-        OutputFormat::parse_raw("human").map(|value| value.to_string()),
-        "human"
-    )]
-    #[case::output_json(
-        OutputFormat::parse_raw("json").map(|value| value.to_string()),
-        "json"
-    )]
-    fn config_enums_round_trip(
-        #[case] parsed: Result<String, &'static [&'static str]>,
-        #[case] expected: &'static str,
-    ) {
-        assert_eq!(parsed.expect("enum value should parse"), expected);
-    }
-
-    #[rstest]
-    #[case::colour(
-        ColourPolicy::from_str("loud").expect_err("invalid colour policy should fail"),
-        &["auto", "always", "never"]
-    )]
-    #[case::spinner(
-        SpinnerMode::from_str("paused").expect_err("invalid spinner mode should fail"),
-        &["enabled", "disabled"]
-    )]
-    #[case::output(
-        OutputFormat::from_str("tap").expect_err("invalid output format should fail"),
-        &["human", "json"]
-    )]
-    fn config_enums_reject_invalid_values(
-        #[case] error: ParseConfigEnumError,
-        #[case] expected_options: &'static [&'static str],
-    ) {
-        assert_eq!(error.valid_options, expected_options);
-    }
-
-    #[test]
-    fn cli_config_alias_resolution_prefers_new_fields() {
-        let config = CliConfig {
-            progress: Some(false),
-            spinner_mode: Some(SpinnerMode::Enabled),
-            diag_json: false,
-            output_format: Some(OutputFormat::Json),
-            ..CliConfig::default()
-        };
-
-        assert!(config.resolved_progress());
-        assert!(config.resolved_diag_json());
-    }
-
-    #[test]
-    fn colour_policy_deserialize_matches_cli_parsing() {
-        for (raw, expected) in [
-            (json!("Always"), ColourPolicy::Always),
-            (json!("ALWAYS"), ColourPolicy::Always),
-            (json!(" always "), ColourPolicy::Always),
-        ] {
-            let parsed: ColourPolicy =
-                serde_json::from_value(raw).expect("colour policy should deserialize");
-            assert_eq!(parsed, expected);
-        }
-    }
-
-    #[test]
-    fn spinner_mode_deserialize_matches_cli_parsing() {
-        for (raw, expected) in [
-            (json!("Enabled"), SpinnerMode::Enabled),
-            (json!("DISABLED"), SpinnerMode::Disabled),
-            (json!(" disabled "), SpinnerMode::Disabled),
-        ] {
-            let parsed: SpinnerMode =
-                serde_json::from_value(raw).expect("spinner mode should deserialize");
-            assert_eq!(parsed, expected);
-        }
-    }
-
-    #[test]
-    fn output_format_deserialize_matches_cli_parsing() {
-        for (raw, expected) in [
-            (json!("Human"), OutputFormat::Human),
-            (json!("JSON"), OutputFormat::Json),
-            (json!(" json "), OutputFormat::Json),
-        ] {
-            let parsed: OutputFormat =
-                serde_json::from_value(raw).expect("output format should deserialize");
-            assert_eq!(parsed, expected);
-        }
-    }
-}
+#[path = "config_tests.rs"]
+mod tests;
