+++ tests/runner_tests.rs
@@ -13,7 +13,10 @@ use test_support::{
     fake_ninja,
 };
 
+#[path = "runner_cases/default_targets.rs"]
+mod default_targets;
 mod fixtures;
+
 use fixtures::create_test_manifest;
 
 /// Fixture: provide a fake `ninja` binary with a configurable exit code.
@@ -41,54 +44,6 @@ fn ninja_in_env() -> Result<(tempfile::TempDir, PathBuf, NinjaEnvGuard)> {
     let env = SystemEnv::new();
     let guard = override_ninja_env(&env, ninja_path.as_path());
     Ok((ninja_dir, ninja_path, guard))
-}
-
-#[cfg(unix)]
-/// Test fixture that installs a temporary fake `ninja` environment.
-///
-/// `_ninja_dir` owns the temporary directory containing the generated fake
-/// `ninja` binary and log file, `_guard` restores the overridden ninja
-/// environment on drop, and `args_log` points to the recorded invocation
-/// arguments emitted by the fake binary.
-struct FakeNinjaFixture {
-    _ninja_dir: tempfile::TempDir,
-    _guard: NinjaEnvGuard,
-    args_log: PathBuf,
-}
-
-#[cfg(unix)]
-#[fixture]
-/// Creates a [`FakeNinjaFixture`] used to simulate ninja behaviour for tests,
-/// returning a temporary environment and the recorded invocation log path.
-fn fake_ninja_fixture() -> Result<FakeNinjaFixture> {
-    use std::fs;
-    use std::os::unix::fs::PermissionsExt;
-
-    let ninja_dir = tempfile::tempdir().context("create fake ninja directory")?;
-    let args_log = ninja_dir.path().join("ninja-args.log");
-    let ninja_path = ninja_dir.path().join("ninja");
-    fs::write(
-        &ninja_path,
-        format!(
-            "#!/bin/sh\nprintf '%s\n' \"$@\" > \"{}\"\nexit 0\n",
-            args_log.display()
-        ),
-    )
-    .with_context(|| format!("write fake ninja script {}", ninja_path.display()))?;
-    let mut permissions = fs::metadata(&ninja_path)
-        .with_context(|| format!("read fake ninja metadata {}", ninja_path.display()))?
-        .permissions();
-    permissions.set_mode(0o755);
-    fs::set_permissions(&ninja_path, permissions)
-        .with_context(|| format!("chmod fake ninja {}", ninja_path.display()))?;
-
-    let env = SystemEnv::new();
-    let guard = override_ninja_env(&env, ninja_path.as_path());
-    Ok(FakeNinjaFixture {
-        _ninja_dir: ninja_dir,
-        _guard: guard,
-        args_log,
-    })
 }
 
 /// Shared setup for tests that rely on `NINJA_ENV`.
@@ -285,37 +240,6 @@ fn run_build_with_emit_creates_parent_dirs() -> Result<()> {
         .context("expected run to succeed with nested emit path")?;
     ensure!(emit_path.exists(), "emit path should be created");
     ensure!(nested_dir.exists(), "nested directory should be created");
-    Ok(())
-}
-
-#[cfg(unix)]
-#[rstest]
-fn run_build_uses_cli_default_targets_when_no_targets_are_requested(
-    fake_ninja_fixture: Result<FakeNinjaFixture>,
-) -> Result<()> {
-    use std::fs;
-
-    let fixture = fake_ninja_fixture?;
-    let (temp, manifest_path) = create_test_manifest()?;
-    let cli = Cli {
-        file: manifest_path,
-        directory: Some(temp.path().to_path_buf()),
-        default_targets: vec![String::from("hello")],
-        command: Some(Commands::Build(BuildArgs {
-            emit: None,
-            targets: Vec::new(),
-        })),
-        ..Cli::default()
-    };
-
-    run(&cli, output_prefs::resolve(None)).context("run build with cli default targets")?;
-
-    let logged_args = fs::read_to_string(&fixture.args_log)
-        .with_context(|| format!("read fake ninja args log {}", fixture.args_log.display()))?;
-    ensure!(
-        logged_args.lines().any(|line| line == "hello"),
-        "expected fake ninja invocation to include default target 'hello', got: {logged_args}"
-    );
     Ok(())
 }
